DeploymentCode/lambdas/classify_spam/requirements.txt
requests
PyJWT
cryptography
httpx